+++ tabbycat/printing/templates/PrintableFeedbackQuestion.vue
@@ -3,7 +3,9 @@
                      question.type === 'tl' ? 'db-flex-item-20' : 'db-flex-item-1']">
     <div class="db-padding-horizontal db-flex-item-1 db-flex-row">
 
-      <div v-if="question.type !== 'tl'" class="db-align-vertical-center db-padding-horizontal db-flex-item db-flex-static db-vertical-center-text">
+      <div v-if="question.type === 'comment'" class="db-align-vertical-center db-flex-item db-vertical-center-text" v-html="question.text"></div>
+
+      <div v-if="question.type !== 'comment' && question.type !== 'tl'" class="db-align-vertical-center db-flex-item db-flex-static db-vertical-center-text">
         {{ question.text }}
       </div>
 
@@ -21,7 +23,7 @@
 
       <div v-if="question.type === 'ss' || question.type === 'ms'" class="db-flex-item db-flex-row">
         <div v-for="choice in question.choice_options" class="db-align-horizontal-center db-align-vertical-center db-flex-item db-center-text db-vertical-center-text">
-          <span class="db-fill-in">{{ choice }}</span>
+          <span class="db-fill-in db-padding-horizontal">{{ choice }}</span>
         </div>
       </div>
 
@@ -31,13 +33,13 @@
         </div>
       </div>
 
-      <div v-if="question.type === 't'" class="db-padding-horizontal db-flex-item-1 db-flex-row db-flex-item db-fill-in"></div>
+      <div v-if="question.type === 't'" class="db-flex-item-1 db-flex-row db-flex-item db-fill-in"></div>
 
       <div v-if="question.type === 'tl'" class="db-flex-column db-flex-item">
-        <div class="db-padding-horizontal db-flex-item db-flex-item-fhs db-align-vertical-center">
+        <div class="db-flex-item db-flex-item-fhs db-align-vertical-center">
           {{ question.text }}
         </div>
-        <div class="db-padding-horizontal db-flex-item db-ruled-bg"></div>
+        <div class="db-flex-item db-ruled-bg"></div>
       </div>
 
     </div>